Submit for GitHub actions tests

+++ controllers/clothingItems.js
@@ -3,7 +3,10 @@ const {
   BAD_REQUEST,
   NOT_FOUND,
   INTERNAL_SERVER_ERROR,
+  CONFLICT,
 } = require("../utils/errors");
+const validator = require("validator");
+const mongoose = require("mongoose");
 
 module.exports.getClothingItems = (req, res) => {
   ClothingItem.find({})
@@ -40,15 +43,32 @@ module.exports.getClothingItem = (req, res) => {
 
 module.exports.createClothingItem = (req, res) => {
   const { name, weather, imageUrl } = req.body;
-  const owner = req.user?._id || "650ffb6de59b1c001b47c9f2";
+  const owner = req.user._id;
+
+  if (!name || !weather || !imageUrl) {
+    return res.status(BAD_REQUEST).send({ message: "All fields are required" });
+  }
+
+  if (typeof name !== "string" || name.length < 2 || name.length > 30) {
+    return res
+      .status(BAD_REQUEST)
+      .send({ message: "Name must be a string between 2 and 30 characters" });
+  }
+
+  if (!validator.isURL(imageUrl)) {
+    return res
+      .status(BAD_REQUEST)
+      .send({ message: "Invalid image URL format" });
+  }
 
   ClothingItem.create({ name, weather, imageUrl, owner })
     .then((item) => res.status(201).send(item))
     .catch((err) => {
+      console.error("❌ Error creating item:", err);
       if (err.name === "ValidationError") {
         return res.status(BAD_REQUEST).send({ message: "Invalid data" });
       }
-      return res
+      res
         .status(INTERNAL_SERVER_ERROR)
         .send({ message: "An error has occurred on the server" });
     });
@@ -57,30 +77,48 @@ module.exports.createClothingItem = (req, res) => {
 module.exports.deleteClothingItem = (req, res) => {
   const { itemId } = req.params;
 
-  ClothingItem.findByIdAndDelete(itemId)
+  ClothingItem.findById(itemId)
     .orFail(() => {
-      const error = new Error("Item not found");
-      error.statusCode = NOT_FOUND;
-      throw error;
+      throw Object.assign(new Error("Item not found"), {
+        statusCode: NOT_FOUND,
+      });
     })
-    .then(() => res.send({ message: "Item deleted" }))
+    .then((item) => {
+      if (!item.owner.equals(req.user._id)) {
+        return res
+          .status(403)
+          .send({ message: "You are not authorized to delete this item" });
+      }
+
+      return ClothingItem.findByIdAndDelete(itemId);
+    })
+    .then((deletedItem) => {
+      if (!deletedItem) {
+        return res.status(NOT_FOUND).send({ message: "Item not found" });
+      }
+      res.status(200).send({ message: "Item deleted successfully" });
+    })
     .catch((err) => {
+      if (err.statusCode === NOT_FOUND) {
+        return res.status(NOT_FOUND).send({ message: err.message });
+      }
       if (err.name === "CastError") {
         return res
           .status(BAD_REQUEST)
           .send({ message: "Invalid item ID format" });
       }
-      if (err.statusCode === NOT_FOUND) {
-        return res.status(NOT_FOUND).send({ message: err.message });
-      }
       return res
         .status(INTERNAL_SERVER_ERROR)
-        .send({ message: "An error has occurred on the server" });
+        .send({ message: "An internal server error occurred" });
     });
 };
 
 module.exports.likeItem = (req, res) => {
   const { itemId } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(itemId)) {
+    return res.status(BAD_REQUEST).send({ message: "Invalid item ID format" });
+  }
 
   ClothingItem.findByIdAndUpdate(
     itemId,
@@ -92,8 +130,14 @@ module.exports.likeItem = (req, res) => {
       error.statusCode = NOT_FOUND;
       throw error;
     })
-    .then((updatedItem) => res.send(updatedItem))
+    .then((updatedItem) => {
+      if (!updatedItem) {
+        return res.status(NOT_FOUND).send({ message: "Item not found" });
+      }
+      res.send(updatedItem);
+    })
     .catch((err) => {
+      console.error("Error liking item:", err);
       if (err.name === "CastError") {
         return res
           .status(BAD_REQUEST)
@@ -104,12 +148,16 @@ module.exports.likeItem = (req, res) => {
       }
       return res
         .status(INTERNAL_SERVER_ERROR)
-        .send({ message: "An error has occurred on the server" });
+        .send({ message: "An internal server error occurred" });
     });
 };
 
 module.exports.dislikeItem = (req, res) => {
   const { itemId } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(itemId)) {
+    return res.status(BAD_REQUEST).send({ message: "Invalid item ID format" });
+  }
 
   ClothingItem.findByIdAndUpdate(
     itemId,
@@ -121,8 +169,14 @@ module.exports.dislikeItem = (req, res) => {
       error.statusCode = NOT_FOUND;
       throw error;
     })
-    .then((updatedItem) => res.send(updatedItem))
+    .then((updatedItem) => {
+      if (!updatedItem) {
+        return res.status(NOT_FOUND).send({ message: "Item not found" });
+      }
+      res.send(updatedItem);
+    })
     .catch((err) => {
+      console.error("Error unliking item:", err);
       if (err.name === "CastError") {
         return res
           .status(BAD_REQUEST)
@@ -133,6 +187,6 @@ module.exports.dislikeItem = (req, res) => {
       }
       return res
         .status(INTERNAL_SERVER_ERROR)
-        .send({ message: "An error has occurred on the server" });
+        .send({ message: "An internal server error occurred" });
     });
 };
